Fixed issue w/ network going away after reboot.
We weren't creating a netplan config and/or and entry in /etc/network/interfaces for br-ex.

Since we don't actually want to touch the host system, I added a wrapper around neutron-server that sets up a transient br-ex before the neutron-server daemon starts. This addresses the issue in a lightweight way that will automatically clean itself up after the snap is uninstalled and the host reboots.