Moved cirros image into $SNAP_COMMON/images
Previously, we were dropping the image into /images. This is not a
great place to put it, as it gets left behind after the snap is
uninstalled, and is generally hidden from the user.

Putting it in $SNAP_COMMON/images makes it easier to find, and means
that it will get cleaned up when the snap is uninstalled.