(WIP) Refactor snap to work with core18.
Giving the snapcraft.yaml a base property helps tremendously with the
efficiency of the build process, and I believe that it puts us in a
better position to reliably support non Ubuntu distros going forward.

This also bases us on long supported bionic libraries, and gives us a
nice place to work from as we add Python 3 and Stein support, as well
as general polish and fixes.

Known issues w/ this WIP PR:
- compiled libvirt looks for iptables in /root/stage/sbin/iptables
  This leads to failures when setting up networks.
  We probably just need to pass the right flag to make config; that
  path is in the staging path for snapcraft.